Seeded the DB, changed student id to decimal for auth0.

+++ models/Student.js
@@ -6,7 +6,7 @@ class Student extends Model {}
 Student.init(
   {
     id: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(24,0),
       allowNull: false,
       primaryKey: true,
       autoIncrement: false,
